change float

+++ src/main/java/tp04/metier/Action.java
@@ -33,7 +33,7 @@ public abstract class Action {
         return libelle;
     }
 
-    public Map<Jour, Integer> getCours() {
+    public Map<Jour, Float> getCours() {
         return this.cours;
     }
